apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
passlib[bcrypt]
bcrypt
python-multipart
pyjwt
